all: finalize ladon and logrus changes
