+++ client/manager_rethinkdb.go
@@ -10,7 +10,7 @@ import (
 	"github.com/ory-am/hydra/pkg"
 	"github.com/pborman/uuid"
 	"github.com/pkg/errors"
-	"golang.org/x/net/context"
+	"context"
 	r "gopkg.in/gorethink/gorethink.v3"
 )
