+++ cmd/server/handler_client_factory.go
@@ -6,7 +6,7 @@ import (
 	"github.com/ory-am/hydra/client"
 	"github.com/ory-am/hydra/config"
 	"github.com/ory/herodot"
-	"golang.org/x/net/context"
+	"context"
 	r "gopkg.in/gorethink/gorethink.v3"
 )
 
@@ -54,7 +54,7 @@ func newClientManager(c *config.Config) client.Manager {
 func newClientHandler(c *config.Config, router *httprouter.Router, manager client.Manager) *client.Handler {
 	ctx := c.Context()
 	h := &client.Handler{
-		H: herodot.NewJSONWriter(c.Context().Logger),
+		H: herodot.NewJSONWriter(c.GetLogger()),
 		W: ctx.Warden, Manager: manager,
 	}
