+++ cmd/server/handler_policy_factory.go
@@ -10,7 +10,7 @@ import (
 func newPolicyHandler(c *config.Config, router *httprouter.Router) *policy.Handler {
 	ctx := c.Context()
 	h := &policy.Handler{
-		H:        herodot.NewJSONWriter(c.Context().Logger),
+		H:        herodot.NewJSONWriter(c.GetLogger()),
 		W:       ctx.Warden,
 		Manager: ctx.LadonManager,
 	}
