+++ oauth2/fosite_store_redis.go
@@ -9,7 +9,7 @@ import (
 	"github.com/ory-am/fosite"
 	"github.com/ory-am/hydra/client"
 	"github.com/pkg/errors"
-	"golang.org/x/net/context"
+	"context"
 	"github.com/go-redis/redis"
 )
