+++ oauth2/fosite_store_rethinkdb.go
@@ -11,7 +11,7 @@ import (
 	"github.com/ory-am/hydra/client"
 	"github.com/ory-am/hydra/pkg"
 	"github.com/pkg/errors"
-	"golang.org/x/net/context"
+	"context"
 	r "gopkg.in/gorethink/gorethink.v3"
 )
