+++ oauth2/introspector_http.go
@@ -5,7 +5,7 @@ import (
 	"encoding/json"
 	"github.com/ory-am/fosite"
 	"github.com/pkg/errors"
-	"golang.org/x/net/context"
+	"context"
 	"golang.org/x/oauth2"
 	"golang.org/x/oauth2/clientcredentials"
 	"io/ioutil"
