+++ oauth2/revocator.go
@@ -1,6 +1,6 @@
 package oauth2
 
-import "golang.org/x/net/context"
+import "context"
 
 type Revocator interface {
 	RevokeToken(ctx context.Context, token string) error
